move Auctioneer frontend to lib.rs, flatten business logic
