+++ crates/astria-auctioneer/src/auctioneer/auction/allocation_rule.rs
